✨ :: implement separate user by team logic
